project/requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==3.2.5
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.4.2
Flask==2.2.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
flatbuffers==22.10.26
fonttools==4.28.5
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.7.0
idna==2.10
importlib-metadata==4.10.0
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
notebook-as-pdf==0.5.0
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.6
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyppeteer==0.2.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
spark-nlp==4.1.0
sparknlp==1.0.0
SQLAlchemy==1.4.44
sqlparse==0.4.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.2.2
whitenoise==6.2.0
widgetsnbextension==3.5.2
wrapt==1.14.1
xgboost==1.6.2
zipp==3.6.0
